--- a/pkg/pipeline/transform/netdb/netdb_test.go
+++ b/pkg/pipeline/transform/netdb/netdb_test.go
@@ -1,0 +1,117 @@
+package netdb
+
+import (
+	"os"
+	"path"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+)
+
+func testingServicesDB() (*ServicesDB, error) {
+	etcProtos, err := os.Open(path.Join("fixtures", "etcProtocols.txt"))
+	if err != nil {
+		return nil, err
+	}
+	defer etcProtos.Close()
+	etcSvcs, err := os.Open(path.Join("fixtures", "etcServices.txt"))
+	if err != nil {
+		return nil, err
+	}
+	defer etcSvcs.Close()
+
+	return LoadServicesDB(etcProtos, etcSvcs)
+}
+
+func TestServicesDB(t *testing.T) {
+	db, err := testingServicesDB()
+	require.NoError(t, err)
+
+	assert.Equal(t, "netbios-dgm", db.ByPortAndProtocolNumber(138, 6))
+	assert.Equal(t, "netbios-dgm", db.ByPortAndProtocolName(138, "tcp"))
+	// verify it also finds service name by protocol alias
+	assert.Equal(t, "netbios-dgm", db.ByPortAndProtocolName(138, "TCP"))
+
+	// verify multiple ports can be associated to the same protocol
+	assert.Equal(t, "ms-sql-s", db.ByPortAndProtocolNumber(1433, 6))
+	assert.Equal(t, "ms-sql-s", db.ByPortAndProtocolName(1433, "tcp"))
+	assert.Equal(t, "ms-sql-s", db.ByPortAndProtocolName(1433, "TCP"))
+	assert.Equal(t, "ms-sql-s", db.ByPortAndProtocolNumber(1433, 17))
+	assert.Equal(t, "ms-sql-s", db.ByPortAndProtocolName(1433, "udp"))
+	assert.Equal(t, "ms-sql-s", db.ByPortAndProtocolName(1433, "UDP"))
+
+	// verify it does search only by port number, if the protocol does not exist
+	assert.Equal(t, "ms-sql-s", db.ByPortAndProtocolNumber(1433, 99999))
+	assert.Equal(t, "ms-sql-s", db.ByPortAndProtocolName(1433, "tralara"))
+
+	// verify it returns nothing if the protocol exist but it's not associated to that port
+	assert.Empty(t, db.ByPortAndProtocolNumber(1433, 18))
+	assert.Empty(t, db.ByPortAndProtocolName(1433, "mux"))
+	assert.Empty(t, db.ByPortAndProtocolName(1433, "MUX"))
+}
+
+func BenchmarkGetProtoByNumber(b *testing.B) {
+	b.StopTimer()
+	db, err := testingServicesDB()
+	if err != nil {
+		b.Fatal(err)
+	}
+	b.StartTimer()
+	for i := 0; i < b.N; i++ {
+		db.ByPortAndProtocolNumber(80, 6)
+		db.ByPortAndProtocolNumber(443, 17)
+		db.ByPortAndProtocolNumber(3306, 17)
+		db.ByPortAndProtocolNumber(27017, 6)
+	}
+}
+
+func BenchmarkGetProtoByName(b *testing.B) {
+	b.StopTimer()
+	db, err := testingServicesDB()
+	if err != nil {
+		b.Fatal(err)
+	}
+	b.StartTimer()
+	for i := 0; i < b.N; i++ {
+		db.ByPortAndProtocolName(80, "tcp")
+		db.ByPortAndProtocolName(443, "udp")
+		db.ByPortAndProtocolName(3306, "UDP")
+		db.ByPortAndProtocolName(27017, "TCP")
+	}
+}
+
+func BenchmarkGetProtoByNumber_Old(b *testing.B) {
+	b.StopTimer()
+	if err := loadOldDB(); err != nil {
+		b.Fatal(err)
+	}
+	b.StartTimer()
+	for i := 0; i < b.N; i++ {
+		GetServByPort(80, GetProtoByNumber(6))
+		GetServByPort(443, GetProtoByNumber(16))
+		GetServByPort(3306, GetProtoByNumber(17))
+		GetServByPort(27017, GetProtoByNumber(6))
+	}
+}
+
+func BenchmarkGetProtoByName_Old(b *testing.B) {
+	b.StopTimer()
+	if err := loadOldDB(); err != nil {
+		b.Fatal(err)
+	}
+	b.StartTimer()
+	for i := 0; i < b.N; i++ {
+		GetServByPort(80, GetProtoByName("tcp"))
+		GetServByPort(443, GetProtoByName("udp"))
+		GetServByPort(3306, GetProtoByName("UDP"))
+		GetServByPort(27017, GetProtoByName("TCP"))
+	}
+}
+
+func loadOldDB() error {
+	return Init(
+		path.Join("fixtures", "etcProtocols.txt"),
+		path.Join("fixtures", "etcServices.txt"),
+	)
+}