--- a/pkg/pipeline/transform/transform_network_test.go
+++ b/pkg/pipeline/transform/transform_network_test.go
@@ -19,6 +19,7 @@ package transform
 
 import (
 	"os"
+	"path"
 	"testing"
 
 	"github.com/netobserv/flowlogs-pipeline/pkg/api"
@@ -122,21 +123,33 @@ func getExpectedOutput() config.GenericMap {
 	}
 }
 
+func getServicesDB(t *testing.T) *netdb.ServicesDB {
+	etcProtos, err := os.Open(path.Join("fixtures", "etcProtocols.txt"))
+	require.NoError(t, err)
+	defer etcProtos.Close()
+	etcSvcs, err := os.Open(path.Join("fixtures", "etcServices.txt"))
+	require.NoError(t, err)
+	defer etcSvcs.Close()
+
+	db, err := netdb.LoadServicesDB(etcProtos, etcSvcs)
+	require.NoError(t, err)
+
+	return db
+}
+
 func Test_Transform(t *testing.T) {
 	entry := test.GetIngestMockEntry(false)
 	rules := getMockNetworkTransformRules()
 	expectedOutput := getExpectedOutput()
 
 	var networkTransform = Network{
-		api.TransformNetwork{
+		TransformNetwork: api.TransformNetwork{
 			Rules: rules,
 		},
+		servicesDB: getServicesDB(t),
 	}
 
 	err := location.InitLocationDB()
-	require.NoError(t, err)
-
-	err = netdb.Init("", "")
 	require.NoError(t, err)
 
 	output := networkTransform.Transform([]config.GenericMap{entry})
@@ -154,15 +167,13 @@ func Test_TransformAddSubnetParseCIDRFailure(t *testing.T) {
 	delete(expectedOutput, "subnet24SrcIP")
 
 	var networkTransform = Network{
-		api.TransformNetwork{
+		TransformNetwork: api.TransformNetwork{
 			Rules: rules,
 		},
+		servicesDB: getServicesDB(t),
 	}
 
 	err := location.InitLocationDB()
-	require.NoError(t, err)
-
-	err = netdb.Init("", "")
 	require.NoError(t, err)
 
 	output := networkTransform.Transform([]config.GenericMap{entry})
@@ -288,7 +299,7 @@ parameters:
 
 func Test_Transform_AddIfScientificNotation(t *testing.T) {
 	newNetworkTransform := Network{
-		api.TransformNetwork{
+		TransformNetwork: api.TransformNetwork{
 			Rules: api.NetworkTransformRules{
 				api.NetworkTransformRule{
 					Input:      "value",
@@ -348,67 +359,4 @@ func Test_Transform_AddIfScientificNotation(t *testing.T) {
 	output = newNetworkTransform.Transform([]config.GenericMap{entry})
 	require.Equal(t, true, output[0]["dir_Evaluate"])
 	require.Equal(t, "out", output[0]["dir"])
-}
-
-func Test_TransformNetworkOperationNameCustomServices(t *testing.T) {
-	rules := api.NetworkTransformRules{
-		api.NetworkTransformRule{
-			Input:      "dstPort",
-			Output:     "service",
-			Type:       "add_service",
-			Parameters: "protocol",
-		},
-	}
-
-	entry := test.GetIngestMockEntry(false)
-
-	var networkTransform = Network{
-		api.TransformNetwork{
-			Rules: rules,
-		},
-	}
-
-	err := netdb.Init("", "")
-	require.NoError(t, err)
-	output := networkTransform.Transform([]config.GenericMap{entry})
-	require.Equal(t, "ssh", output[0]["service"])
-
-	// check custom services file
-	customService := "custom_ssh             22/tcp                          # The Secure Shell (SSH) Protocol\n"
-	err = os.WriteFile("/tmp/services", []byte(customService), 0644)
-	require.NoError(t, err)
-
-	err = netdb.Init("", "/tmp/services")
-	require.NoError(t, err)
-	output = networkTransform.Transform([]config.GenericMap{entry})
-	require.Equal(t, "custom_ssh", output[0]["service"])
-
-	rules = api.NetworkTransformRules{
-		api.NetworkTransformRule{
-			Input:      "dstPort",
-			Output:     "service_protocol_num",
-			Type:       "add_service",
-			Parameters: "protocol_num",
-		},
-	}
-
-	networkTransform = Network{
-		api.TransformNetwork{
-			Rules: rules,
-		},
-	}
-
-	// check custom protocol file
-	customProtocol := "custom_tcp     6       CUSTOM_TCP             # transmission control protocol\n"
-	err = os.WriteFile("/tmp/protocols", []byte(customProtocol), 0644)
-	require.NoError(t, err)
-
-	customService = "custom_ssh             22/custom_TCP                          # The Secure Shell (SSH) Protocol\n"
-	err = os.WriteFile("/tmp/services", []byte(customService), 0644)
-	require.NoError(t, err)
-
-	err = netdb.Init("/tmp/protocols", "/tmp/services")
-	require.NoError(t, err)
-	output = networkTransform.Transform([]config.GenericMap{entry})
-	require.Equal(t, "custom_ssh", output[0]["service_protocol_num"])
 }