… pkg/pipeline/transform/netdb/fixtures/etcServices.txt