Benchmarks comparing new vs old version:

```
BenchmarkGetProtoByNumber-12            162093156               74.11 ns/op
BenchmarkGetProtoByName-12              93009231               126.4 ns/op

BenchmarkGetProtoByNumber_Old-12          945291             12846 ns/op
BenchmarkGetProtoByName_Old-12          17034988               706.8 ns/op
```

To reproduce all the benchmarks, you should checkout commit `5ac2502081c1a47101f758030b988466cbf554ec`, as the old code is removed in the head commit.